+++ src/path.rs
@@ -1,7 +1,10 @@
 use derive_more::From;
-use orion_error::{ErrorCode, ErrorOwe, ErrorWith, StructError, ToStructError, UvsFrom, UvsReason};
+use orion_error::compat_traits::ErrorOweSource;
+use orion_error::traits_ext::ToStructError;
+use orion_error::{DomainReason, ErrorCode, ErrorWith, StructError, UvsFrom, UvsReason};
 use std::{fs, path::Path};
 use thiserror::Error;
+
 #[derive(Clone, Debug, PartialEq, Error, From)]
 pub enum PathReason {
     #[error("brief {0}")]
@@ -9,6 +12,8 @@ pub enum PathReason {
     #[error("{0}")]
     Uvs(UvsReason),
 }
+
+impl DomainReason for PathReason {}
 
 impl ErrorCode for PathReason {
     fn error_code(&self) -> i32 {
@@ -21,18 +26,24 @@ impl ErrorCode for PathReason {
 
 pub type PathResult<T> = Result<T, StructError<PathReason>>;
 pub type PathError = StructError<PathReason>;
+
 pub fn make_clean_path(path: &Path) -> PathResult<()> {
     if path.exists() {
-        std::fs::remove_dir_all(path).owe_sys().with(path)?;
+        std::fs::remove_dir_all(path)
+            .owe_sys_source()
+            .with_context(path)?;
     }
-    std::fs::create_dir_all(path).owe_sys().with(path)?;
+    std::fs::create_dir_all(path)
+        .owe_sys_source()
+        .with_context(path)?;
     Ok(())
 }
+
 pub fn ensure_path<P: AsRef<Path>>(path: P) -> PathResult<P> {
     if !path.as_ref().exists() {
         std::fs::create_dir_all(path.as_ref())
-            .owe_sys()
-            .with(path.as_ref())?;
+            .owe_sys_source()
+            .with_context(path.as_ref())?;
     }
     Ok(path)
 }
@@ -41,21 +52,21 @@ pub fn make_new_path(path: &Path) -> PathResult<()> {
     if path.exists() {
         return PathReason::from_res()
             .err_result()
-            .want("path exists")
-            .with(path);
+            .doing("path exists")
+            .with_context(path);
     }
-    std::fs::create_dir_all(path).owe_sys()?;
+    std::fs::create_dir_all(path).owe_sys_source()?;
     Ok(())
 }
 
 pub fn get_sub_dirs(path: &Path) -> PathResult<Vec<std::path::PathBuf>> {
     let mut dirs = Vec::new();
     for entry in fs::read_dir(path)
-        .owe_res()
-        .with(path)
-        .want("read sub dirs")?
+        .owe_res_source()
+        .with_context(path)
+        .doing("read sub dirs")?
     {
-        let entry = entry.owe_res()?;
+        let entry = entry.owe_res_source()?;
         let path = entry.path();
         if path.is_dir() {
             dirs.push(path);